requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
cloud-sql-python-connector[pg8000]==1.4.3
pg8000==1.30.3

# Scheduler
APScheduler==3.10.4

# Production Server
gunicorn==21.2.0

# Environment
python-dotenv==1.0.0

# Date/Time
pytz==2024.1
